+++ db/migrate/20160927120513_make_member_counts_unique.rb
@@ -1,0 +1,15 @@
+# encoding: utf-8
+
+#  Copyright (c) 2012-2014, Pfadibewegung Schweiz. This file is part of
+#  hitobito_pbs and licensed under the Affero General Public License version 3
+#  or later. See the COPYING file at the top-level directory or at
+#  https://github.com/hitobito/hitobito_pbs.
+
+class MakeMemberCountsUnique < ActiveRecord::Migration
+  def change
+    add_index :member_counts,
+              [:year, :flock_id, :born_in],
+              unique: true,
+              name: 'index_member_counts_unique_per_year'
+  end
+end