+++ spec/domain/export/tabular/events/bsv_list_spec.rb
@@ -7,11 +7,11 @@
 
 require 'spec_helper'
 
-describe Export::Csv::Events::BsvList do
+describe Export::Tabular::Events::BsvList do
   let(:course) { events(:top_course) }
 
   context 'export' do
-    let(:lines) { Export::Csv::Events::BsvList.export([course, course]).split("\r\n") }
+    let(:lines) { Export::Tabular::Events::BsvList.csv([course, course]).split("\r\n") }
     let(:headers) {  lines.first.encode('UTF-8').split(';') }
 
     it 'exports headers' do