+++ CHANGELOG.md
@@ -1,7 +1,7 @@
-# Hitobito PBS Changelog
+# Hitobito Jubla Changelog
 
 ## Version 1.10
 
 * Scharnamen enthalten bei der Anzeige von Rollen immer auch die Art (Jungwacht/Blauring/Jubla).
-* Neues Feld `Jubla Sachversicherung` für Scharen. 
+* Neues Feld `Jubla Sachversicherung` für Scharen.
 * Feld `Jubla Versicherung` auf Schar wurde umbenannt nach `Jubla Haftpflicht`.